refactor: create enums package inside the domain layer
