+++ src/main/java/io/pakland/mdas/githubstats/application/dto/PullRequestDTO.java
@@ -1,6 +1,6 @@
 package io.pakland.mdas.githubstats.application.dto;
 
-import io.pakland.mdas.githubstats.domain.entity.PullRequestState;
+import io.pakland.mdas.githubstats.domain.enums.PullRequestState;
 
 public interface PullRequestDTO {
     Integer getId();
