+++ src/main/java/io/pakland/mdas/githubstats/domain/enums/EntityType.java
@@ -1,4 +1,4 @@
-package io.pakland.mdas.githubstats.domain;
+package io.pakland.mdas.githubstats.domain.enums;
 
 public enum EntityType {
     ORGANIZATION,
